+++ R/Driver.R
@@ -61,6 +61,8 @@ setMethod(
 #' name for the OdbcConnect object returned from `dbConnect()`. However if the
 #' driver does not return a valid value it can be set manually with this
 #' parameter.
+#' @param attributes An S4 object of connection attributes that are passed
+#' prior to the connection being established.  See \link{ConnectionAttributes}.
 #' @param ... Additional ODBC keywords, these will be joined with the other
 #' arguments to form the final connection string.
 #' @param .connection_string A complete connection string, useful if you are
@@ -96,6 +98,7 @@ setMethod(
     uid = NULL,
     pwd = NULL,
     dbms.name = NULL,
+    attributes = NULL,
     .connection_string = NULL) {
 
     con <- OdbcConnection(
@@ -112,6 +115,7 @@ setMethod(
       uid = uid,
       pwd = pwd,
       dbms.name = dbms.name,
+      attributes = attributes,
       .connection_string = .connection_string)
 
     # perform the connection notification at the top level, to ensure that it's had